requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
python-dotenv==1.0.1
langchain==0.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.2
authlib
requests
numpy